Add take_or_clone_val.
When one is popping back up a cactus stack, it's common to want access to a
node's data just before it is dropped. Via the val() method one has no choice
but to clone() data. However, if the node has no children one can safely consume
the node and move the val data to the caller. This method puts this under one
hood: if possible the val data is moved, otherwise it is cloned.