Use the precedent of Rc::try_unwrap instead of the custom "take_or_clone_val".
take_or_clone_val has two problems: 1) if you don't want to clone the value
if the take fails, tough luck 2) it's a non-standard API.

This commit tentatively remodels the API along the lines of the standard
library's Rc::try_unwrap. This has the virtue of familiarity and flexibility,
but if you want to clone it's more of a pain. This:

  let x = cactus.take_or_clone_val().unwrap();

becomes:

  let x = cactus.try_unwrap().unwrap_or_else(|c| c.val().unwrap().clone())